refactored test structure to separate groups of tests
