--- src/test/java/com/dataart/qa/workshop/categories/RegressionTests.java
@@ -1,4 +1,0 @@
-package com.dataart.qa.workshop.categories;
-
-public interface RegressionTests {
-}
