+++ src/test/java/com/dataart/qa/workshop/grouping/categories/RegressionTests.java
@@ -1,0 +1,4 @@
+package com.dataart.qa.workshop.grouping.categories;
+
+public interface RegressionTests {
+}
