+++ src/test/java/com/dataart/qa/workshop/suites/MySuite.java
@@ -1,11 +1,11 @@
 package com.dataart.qa.workshop.suites;
 
 import com.dataart.qa.workshop.test.AssertTest;
-import com.dataart.qa.workshop.test.Example1Test;
+import com.dataart.qa.workshop.test.MyRunnerTest;
 import org.junit.runner.RunWith;
 import org.junit.runners.Suite;
 
 @RunWith(Suite.class)
-@Suite.SuiteClasses({AssertTest.class, Example1Test.class})
+@Suite.SuiteClasses({AssertTest.class, MyRunnerTest.class})
 public class MySuite {
 }
