+++ src/test/java/com/dataart/qa/workshop/test/MyRunnerTest.java
@@ -1,0 +1,22 @@
+package com.dataart.qa.workshop.test;
+
+import com.dataart.qa.workshop.junit.runners.MyRunner;
+import org.junit.Assert;
+import org.junit.Test;
+import org.junit.runner.RunWith;
+
+@RunWith(MyRunner.class)
+public class MyRunnerTest {
+    private static String actual = "Actual string";
+    private static String expected = "Actual string";// "Expected string"
+
+    @Test
+    public void example1_the_first_test() {
+        Assert.assertEquals("Wrong string found", expected, actual);
+    }
+
+    @Test
+    public void example1_the_second_test() {
+        Assert.assertEquals("Wrong string found", expected, actual);
+    }
+}
